Refactor usage of OnBeforeSend; detect SDK related error in the error message.
