+++ uikit/src/main/java/com/midtrans/sdk/uikit/activities/PaymentMethodsActivity.java
@@ -4,7 +4,6 @@ import android.content.DialogInterface;
 import android.content.Intent;
 import android.graphics.PorterDuff;
 import android.graphics.drawable.Drawable;
-import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
 import android.support.design.widget.AppBarLayout;
@@ -23,22 +22,16 @@ import android.widget.Toast;
 
 import com.koushikdutta.ion.Ion;
 import com.midtrans.raygun.RaygunClient;
-import com.midtrans.raygun.RaygunOnBeforeSend;
-import com.midtrans.raygun.messages.RaygunErrorStackTraceLineMessage;
-import com.midtrans.raygun.messages.RaygunMessage;
-import com.midtrans.raygun.messages.RaygunMessageDetails;
 import com.midtrans.sdk.corekit.callback.CheckoutCallback;
 import com.midtrans.sdk.corekit.callback.TransactionOptionsCallback;
 import com.midtrans.sdk.corekit.core.Constants;
 import com.midtrans.sdk.corekit.core.Logger;
 import com.midtrans.sdk.corekit.core.MidtransSDK;
-import com.midtrans.sdk.corekit.core.SdkUtil;
 import com.midtrans.sdk.corekit.core.TransactionRequest;
 import com.midtrans.sdk.corekit.core.themes.ColorTheme;
 import com.midtrans.sdk.corekit.core.themes.CustomColorTheme;
 import com.midtrans.sdk.corekit.models.CustomerDetails;
 import com.midtrans.sdk.corekit.models.ItemDetails;
-import com.midtrans.sdk.corekit.models.MerchantPreferences;
 import com.midtrans.sdk.corekit.models.PaymentDetails;
 import com.midtrans.sdk.corekit.models.PaymentMethodsModel;
 import com.midtrans.sdk.corekit.models.TransactionResponse;
@@ -53,12 +46,12 @@ import com.midtrans.sdk.corekit.utilities.Utils;
 import com.midtrans.sdk.uikit.BuildConfig;
 import com.midtrans.sdk.uikit.PaymentMethods;
 import com.midtrans.sdk.uikit.R;
+import com.midtrans.sdk.uikit.UiKitOnBeforeSend;
 import com.midtrans.sdk.uikit.adapters.ItemDetailsAdapter;
 import com.midtrans.sdk.uikit.adapters.PaymentMethodsAdapter;
 import com.midtrans.sdk.uikit.models.EnabledPayments;
 import com.midtrans.sdk.uikit.models.ItemViewDetails;
 import com.midtrans.sdk.uikit.models.MessageInfo;
-import com.midtrans.sdk.uikit.utilities.DeviceUtils;
 import com.midtrans.sdk.uikit.utilities.MessageUtil;
 import com.midtrans.sdk.uikit.utilities.SdkUIFlowUtil;
 import com.midtrans.sdk.uikit.utilities.UiKitConstants;
@@ -477,57 +470,7 @@ public class PaymentMethodsActivity extends BaseActivity implements PaymentMetho
     private void initCustomTrackingProperties() {
 
         if (BuildConfig.FLAVOR.equals(UiKitConstants.ENVIRONMENT_PRODUCTION)) {
-            RaygunClient.setOnBeforeSend(new RaygunOnBeforeSend() {
-                @Override
-                public RaygunMessage onBeforeSend(RaygunMessage raygunMessage) {
-
-                    RaygunErrorStackTraceLineMessage[] stackTrace = raygunMessage.getDetails().getError().getStackTrace();
-                    if (stackTrace != null) {
-                        for (RaygunErrorStackTraceLineMessage message : stackTrace) {
-                            String className = message.getClassName();
-                            if (className.contains("com.midtrans")) {
-                                try {
-                                    RaygunMessageDetails details = raygunMessage.getDetails();
-                                    details.setGroupingKey(UiKitConstants.KEY_TRACKING_GROUP);
-
-                                    Map<String, String> map = new HashMap<>();
-
-                                    MerchantPreferences preferences = MidtransSDK.getInstance().getMerchantData().getPreference();
-
-                                    if (preferences != null) {
-                                        map.put(UiKitConstants.KEY_TRACKING_MERCHANT_NAME, preferences.getDisplayName());
-                                    }
-                                    String[] appInfo = DeviceUtils.getApplicationName(PaymentMethodsActivity.this);
-                                    map.put(UiKitConstants.KEY_TRACKING_HOST_APP, appInfo[0]);
-                                    map.put(UiKitConstants.KEY_TRACKING_HOST_APP_VERSION, appInfo[1]);
-                                    map.put(UiKitConstants.KEY_TRACKING_DEVICE_ID, SdkUtil.getDeviceId(PaymentMethodsActivity.this));
-                                    map.put(UiKitConstants.KEY_TRACKING_LANGUAGE, Locale.getDefault().getLanguage());
-                                    map.put(UiKitConstants.KEY_TRACKING_DEVICE_MODEL, Build.MODEL);
-                                    map.put(UiKitConstants.KEY_TRACKING_DEVICE_TYPE, Build.BRAND);
-                                    map.put(UiKitConstants.KEY_TRACKING_TIME_STAMP, String.valueOf(System.currentTimeMillis()));
-                                    map.put(UiKitConstants.KEY_TRACKING_NETWORK, DeviceUtils.getConnectivityType(PaymentMethodsActivity.this));
-                                    map.put(UiKitConstants.KEY_TRACKING_OS_VERSION, String.valueOf(Build.VERSION.SDK_INT));
-                                    map.put(UiKitConstants.KEY_TRACKING_PLATFORM, UiKitConstants.VALUE_TRACKING_PLATFORM);
-                                    map.put(UiKitConstants.KEY_TRACKING_SCREEN_SIZE, DeviceUtils.getDisplaySize(PaymentMethodsActivity.this));
-                                    map.put(UiKitConstants.KEY_TRACKING_SDK_VERSION, BuildConfig.VERSION_NAME);
-                                    map.put(UiKitConstants.KEY_TRACKING_CPU_USAGE, DeviceUtils.getTotalCpuUsage());
-                                    map.put(UiKitConstants.KEY_TRACKING_MEMORY_USAGE, DeviceUtils.getMemoryUsage());
-                                    map.put(UiKitConstants.KEY_TRACKING_ENVIRONMENT, BuildConfig.FLAVOR);
-
-                                    details.setUserCustomData(map);
-
-                                } catch (Exception e) {
-                                    Logger.d(TAG, "raygun:" + e.getMessage());
-                                }
-
-                                return raygunMessage;
-                            }
-                        }
-                    }
-                    return null;
-
-                }
-            });
+            RaygunClient.setOnBeforeSend(new UiKitOnBeforeSend(this));
         }
     }
